Cleaned up codebase. Fixed package.json. Organized project
