+++ package.json
@@ -1,17 +1,17 @@
 {
-  "name": "javascript-action",
+  "name": "github-user-audit",
   "version": "1.0.0",
-  "description": "JavaScript Action Template",
+  "description": "Uses GitHub GraphQL to audit users permissions on an organization's repositories. Generates a summary and a csv file.",
   "main": "index.js",
   "scripts": {
     "lint": "eslint .",
-    "prepare": "ncc build index.js -o dist --source-map --license licenses.txt",
+    "prepare": "ncc build src/index.js -o dist --source-map --license licenses.txt",
     "test": "cross-env NODE_OPTIONS=--experimental-vm-modules jest",
     "all": "npm run lint && npm run prepare && npm run test"
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/actions/javascript-action.git"
+    "url": "git@github.com:gh-customer-success/github-user-audit.git"
   },
   "keywords": [
     "GitHub",
@@ -21,14 +21,13 @@
   "author": "",
   "license": "MIT",
   "bugs": {
-    "url": "https://github.com/actions/javascript-action/issues"
+    "url": "https://github.com/gh-customer-success/github-user-audit/issues"
   },
-  "homepage": "https://github.com/actions/javascript-action#readme",
+  "homepage": "https://github.com/gh-customer-success/github-user-audit#readme",
   "dependencies": {
     "@actions/artifact": "^1.1.1",
     "@actions/core": "^1.10.0",
     "@octokit/graphql": "^5.0.5",
-    "bottleneck": "^2.19.5",
     "csv-stringify": "^6.3.0",
     "octokit": "^2.0.14"
   },
