+++ src/index.js
@@ -1,10 +1,6 @@
 import core from '@actions/core';
-import { graphql } from '@octokit/graphql';
-import * as fs from 'fs';
-import path from 'path';
-const query = fs.readFileSync(path.join(__dirname, './audit.gql'), 'utf8');
-import { getTeams, getRepos } from './sort-audit.js';
-import { teamsCSV, repoCSV } from './create-csv.js';
+import { getRepos } from './sort-audit.js';
+import { repoCSV } from './create-csv.js';
 import { getAllData } from './get-repository-collaborators.js';
 import { generatePermissionsCount } from './step-summary-table.js';
 // most @actions toolkit packages have async methods
@@ -15,9 +11,8 @@ async function run() {
     const owner = core.getInput('owner');
     const token = core.getInput('api_token');
 
-
-
     //call the getTeams function and pass in data
+    //not implemented for now
     // const teams = getTeams(data); 
     //call the getRepos function and pass in data 
 
@@ -26,6 +21,7 @@ async function run() {
 
 
       //create the csv and upload it as an artifact
+      //not implemented for now
       // teamsCSV(teams);
       const repos = getRepos(repositories);
       repoCSV(repos);
@@ -33,7 +29,7 @@ async function run() {
       core.setOutput('repos', JSON.stringify(repositories));
       //run an echo command to pipe the output of generateOutputString to $GITHUB_STEP_SUMMARY
       generatePermissionsCount(repositories);
-      
+
     }).catch((error) => {
       console.error(error);
       core.setFailed(error.message);
